refactor: apply lazy loading to remaining CLI commands
- Refactored auth, config, delete, export, import, validate, debug, and modelScan commands
- Moved heavy imports from command registration to action files
- Maintained backward compatibility with re-exports where needed
- Removed benchmark scripts that were causing linting issues
- All commands now follow the same lazy loading pattern

This completes the lazy loading refactoring for all CLI commands, ensuring
consistent minimal imports during command registration phase.
