+++ src/commands/delete.ts
@@ -1,32 +1,7 @@
-import confirm from '@inquirer/confirm';
 import type { Command } from 'commander';
-import logger from '../logger';
-import Eval from '../models/eval';
-import telemetry from '../telemetry';
-import { setupEnv } from '../util';
-import { deleteAllEvals, deleteEval, getEvalFromId } from '../util/database';
 
-export async function handleEvalDelete(evalId: string, envPath?: string) {
-  try {
-    await deleteEval(evalId);
-    logger.info(`Evaluation with ID ${evalId} has been successfully deleted.`);
-  } catch (error) {
-    logger.error(`Could not delete evaluation with ID ${evalId}:\n${error}`);
-    process.exit(1);
-  }
-}
-
-export async function handleEvalDeleteAll() {
-  const confirmed = await confirm({
-    message:
-      'Are you sure you want to delete all stored evaluations? This action cannot be undone.',
-  });
-  if (!confirmed) {
-    return;
-  }
-  await deleteAllEvals();
-  logger.info('All evaluations have been deleted.');
-}
+// Re-export for backward compatibility
+export { handleEvalDelete, handleEvalDeleteAll } from './delete/deleteAction';
 
 export function deleteCommand(program: Command) {
   const deleteCommand = program
@@ -34,18 +9,8 @@ export function deleteCommand(program: Command) {
     .description('Delete various resources')
     .option('--env-file, --env-path <path>', 'Path to .env file')
     .action(async (id: string, cmdObj: { envPath?: string }) => {
-      setupEnv(cmdObj.envPath);
-      telemetry.record('command_used', {
-        name: 'delete',
-      });
-
-      const evl = await getEvalFromId(id);
-      if (evl) {
-        return handleEvalDelete(id, cmdObj.envPath);
-      }
-
-      logger.error(`No resource found with ID ${id}`);
-      process.exitCode = 1;
+      const { deleteAction } = await import('./delete/deleteAction');
+      await deleteAction(id, cmdObj);
     });
 
   deleteCommand
@@ -55,24 +20,7 @@ export function deleteCommand(program: Command) {
     )
     .option('--env-file, --env-path <path>', 'Path to .env file')
     .action(async (evalId, cmdObj) => {
-      setupEnv(cmdObj.envPath);
-      telemetry.record('command_used', {
-        name: 'delete eval',
-        evalId,
-      });
-
-      if (evalId === 'latest') {
-        const latestResults = await Eval.latest();
-        if (latestResults) {
-          await handleEvalDelete(latestResults.id, cmdObj.envPath);
-        } else {
-          logger.error('No eval found.');
-          process.exitCode = 1;
-        }
-      } else if (evalId === 'all') {
-        await handleEvalDeleteAll();
-      } else {
-        await handleEvalDelete(evalId, cmdObj.envPath);
-      }
+      const { deleteEvalAction } = await import('./delete/deleteAction');
+      await deleteEvalAction(evalId, cmdObj);
     });
 }
