+++ src/commands/modelScan.ts
@@ -1,20 +1,7 @@
-import chalk from 'chalk';
-import { spawn } from 'child_process';
-import { exec } from 'child_process';
 import type { Command } from 'commander';
-import { promisify } from 'util';
-import logger from '../logger';
 
-const execAsync = promisify(exec);
-
-export async function checkModelAuditInstalled(): Promise<boolean> {
-  try {
-    await execAsync('which modelaudit');
-    return true;
-  } catch {
-    return false;
-  }
-}
+// Re-export for backward compatibility
+export { checkModelAuditInstalled } from './modelScan/modelScanAction';
 
 export function modelScanCommand(program: Command): void {
   program
@@ -39,73 +26,7 @@ export function modelScanCommand(program: Command): void {
     .option('--max-file-size <bytes>', 'Maximum file size to scan in bytes')
     .option('--max-total-size <bytes>', 'Maximum total bytes to scan before stopping')
     .action(async (paths: string[], options) => {
-      if (!paths || paths.length === 0) {
-        logger.error(
-          'No paths specified. Please provide at least one model file or directory to scan.',
-        );
-        process.exit(1);
-      }
-
-      // Check if modelaudit is installed
-      const isModelAuditInstalled = await checkModelAuditInstalled();
-      if (!isModelAuditInstalled) {
-        logger.error('ModelAudit is not installed.');
-        logger.info(`Please install it using: ${chalk.green('pip install modelaudit')}`);
-        logger.info('For more information, visit: https://www.promptfoo.dev/docs/model-audit/');
-        process.exit(1);
-      }
-
-      const args = ['scan', ...paths];
-
-      // Add options
-      if (options.blacklist && options.blacklist.length > 0) {
-        options.blacklist.forEach((pattern: string) => {
-          args.push('--blacklist', pattern);
-        });
-      }
-
-      if (options.format) {
-        args.push('--format', options.format);
-      }
-
-      if (options.output) {
-        args.push('--output', options.output);
-      }
-
-      if (options.timeout) {
-        args.push('--timeout', options.timeout.toString());
-      }
-
-      if (options.verbose) {
-        args.push('--verbose');
-      }
-
-      if (options.maxFileSize) {
-        args.push('--max-file-size', options.maxFileSize);
-      }
-
-      if (options.maxTotalSize) {
-        args.push('--max-total-size', options.maxTotalSize);
-      }
-
-      logger.info(`Running model scan on: ${paths.join(', ')}`);
-
-      const modelAudit = spawn('modelaudit', args, { stdio: 'inherit' });
-
-      modelAudit.on('error', (error) => {
-        logger.error(`Failed to start modelaudit: ${error.message}`);
-        logger.info('Make sure modelaudit is installed and available in your PATH.');
-        logger.info('Install it using: pip install modelaudit');
-        process.exit(1);
-      });
-
-      modelAudit.on('close', (code) => {
-        if (code === 0) {
-          logger.info('Model scan completed successfully.');
-        } else {
-          logger.error(`Model scan completed with issues. Exit code: ${code}`);
-          process.exit(code || 1);
-        }
-      });
+      const { modelScanAction } = await import('./modelScan/modelScanAction');
+      await modelScanAction(paths, options);
     });
 }
