+++ src/commands/validate.ts
@@ -1,53 +1,13 @@
-import chalk from 'chalk';
 import type { Command } from 'commander';
-import dedent from 'dedent';
-import { fromError } from 'zod-validation-error';
-import logger from '../logger';
-import telemetry from '../telemetry';
 import type { UnifiedConfig } from '../types';
-import { TestSuiteSchema, UnifiedConfigSchema } from '../types';
-import { setupEnv } from '../util';
-import { resolveConfigs } from '../util/config/load';
 
 interface ValidateOptions {
   config?: string[];
   envPath?: string;
 }
 
-export async function doValidate(
-  opts: ValidateOptions,
-  defaultConfig: Partial<UnifiedConfig>,
-  defaultConfigPath: string | undefined,
-): Promise<void> {
-  setupEnv(opts.envPath);
-  const configPaths = opts.config || (defaultConfigPath ? [defaultConfigPath] : undefined);
-  try {
-    const { config, testSuite } = await resolveConfigs(
-      { ...opts, config: configPaths },
-      defaultConfig,
-    );
-    const configParse = UnifiedConfigSchema.safeParse(config);
-    if (!configParse.success) {
-      logger.error(
-        dedent`Configuration validation error:
-Config file path(s): ${Array.isArray(configPaths) ? configPaths.join(', ') : (configPaths ?? 'N/A')}
-${fromError(configParse.error).message}`,
-      );
-      process.exitCode = 1;
-      return;
-    }
-    const suiteParse = TestSuiteSchema.safeParse(testSuite);
-    if (!suiteParse.success) {
-      logger.error(dedent`Test suite validation error:\n${fromError(suiteParse.error).message}`);
-      process.exitCode = 1;
-      return;
-    }
-    logger.info(chalk.green('Configuration is valid.'));
-  } catch (err) {
-    logger.error(`Failed to validate configuration: ${err instanceof Error ? err.message : err}`);
-    process.exitCode = 1;
-  }
-}
+// Re-export for backward compatibility
+export { doValidate } from './validate/validateAction';
 
 export function validateCommand(
   program: Command,
@@ -62,7 +22,7 @@ export function validateCommand(
       'Path to configuration file. Automatically loads promptfooconfig.yaml',
     )
     .action(async (opts: ValidateOptions) => {
-      telemetry.record('command_used', { name: 'validate' });
-      await doValidate(opts, defaultConfig, defaultConfigPath);
+      const { validateAction } = await import('./validate/validateAction');
+      await validateAction(opts, defaultConfig, defaultConfigPath);
     });
 }
